Draft

+++ Tests/TestDoFs.cpp
@@ -52,11 +52,11 @@ TEST(Discrete, DoF_var1)
     size_t varId = dofManagerA.addVariableDoFs(0, 0, 10);
     dofManagerA.construct();
 
-    ASSERT_EQ(1, dofManagerA.getNumberOfVariables());
-    ASSERT_EQ(1, dofManagerA.getNumberOfMeshes());
+    ASSERT_EQ(1u, dofManagerA.getNumberOfVariables());
+    ASSERT_EQ(1u, dofManagerA.getNumberOfMeshes());
     //ASSERT_EQ(dofMap1->getNumberOfDiscretePoints(), 10);
-    ASSERT_EQ(0, dofManagerA.mapEqsID(varId, 0, 0));
-    ASSERT_EQ(9, dofManagerA.mapEqsID(varId, 0, 9));
+    ASSERT_EQ(0u, dofManagerA.mapEqsID(varId, 0, 0));
+    ASSERT_EQ(9u, dofManagerA.mapEqsID(varId, 0, 9));
 };
 
 TEST(Discrete, DoF_var2_byDof)
@@ -66,12 +66,12 @@ TEST(Discrete, DoF_var2_byDof)
     size_t dofIdB2 = dofManagerB.addVariableDoFs(0, 0, 10);
     dofManagerB.setNumberingType(DofNumberingType::BY_VARIABLE);
     dofManagerB.construct();
-    ASSERT_EQ(2, dofManagerB.getNumberOfVariables());
+    ASSERT_EQ(2u, dofManagerB.getNumberOfVariables());
     //ASSERT_EQ(dofManagerB.getTotalNumberOfDiscretePoints(), 20);
-    ASSERT_EQ(0, dofManagerB.mapEqsID(dofIdB1, 0, 0));
-    ASSERT_EQ(9, dofManagerB.mapEqsID(dofIdB1, 0, 9));
-    ASSERT_EQ(10, dofManagerB.mapEqsID(dofIdB2, 0, 0));
-    ASSERT_EQ(19, dofManagerB.mapEqsID(dofIdB2, 0, 9));
+    ASSERT_EQ(0u, dofManagerB.mapEqsID(dofIdB1, 0, 0));
+    ASSERT_EQ(9u, dofManagerB.mapEqsID(dofIdB1, 0, 9));
+    ASSERT_EQ(10u, dofManagerB.mapEqsID(dofIdB2, 0, 0));
+    ASSERT_EQ(19u, dofManagerB.mapEqsID(dofIdB2, 0, 9));
 };
 
 TEST(Discrete, DoF_var2_byPoint)
@@ -81,12 +81,12 @@ TEST(Discrete, DoF_var2_byPoint)
     size_t dofIdB2 = dofManagerB.addVariableDoFs(0, 0, 10);
     dofManagerB.setNumberingType(DofNumberingType::BY_POINT);
     dofManagerB.construct();
-    ASSERT_EQ(2, dofManagerB.getNumberOfVariables());
-    ASSERT_EQ(1, dofManagerB.getNumberOfMeshes());
-    ASSERT_EQ(0, dofManagerB.mapEqsID(dofIdB1, 0, 0));
-    ASSERT_EQ(18, dofManagerB.mapEqsID(dofIdB1, 0, 9));
-    ASSERT_EQ(1, dofManagerB.mapEqsID(dofIdB2, 0, 0));
-    ASSERT_EQ(19, dofManagerB.mapEqsID(dofIdB2, 0, 9));
+    ASSERT_EQ(2u, dofManagerB.getNumberOfVariables());
+    ASSERT_EQ(1u, dofManagerB.getNumberOfMeshes());
+    ASSERT_EQ(0u, dofManagerB.mapEqsID(dofIdB1, 0, 0));
+    ASSERT_EQ(18u, dofManagerB.mapEqsID(dofIdB1, 0, 9));
+    ASSERT_EQ(1u, dofManagerB.mapEqsID(dofIdB2, 0, 0));
+    ASSERT_EQ(19u, dofManagerB.mapEqsID(dofIdB2, 0, 9));
 }
 
 TEST(Discrete, DoF_inactive)
@@ -97,11 +97,11 @@ TEST(Discrete, DoF_inactive)
     dofManager->setNumberingType(DofNumberingType::BY_POINT);
     dofManager->construct();
 
-    ASSERT_EQ(1, dofManager->getNumberOfVariables());
-    ASSERT_EQ(8, dofManager->getTotalNumberOfActiveDoFs());
-    ASSERT_EQ(8, dofManager->getTotalNumberOfActiveDoFsWithoutGhost());
-    ASSERT_EQ(0, dofManager->mapEqsID(varId, 0, 0));
-    ASSERT_EQ(-1, dofManager->mapEqsID(varId, 0, 8));
+    ASSERT_EQ(1u, dofManager->getNumberOfVariables());
+    ASSERT_EQ(8u, dofManager->getTotalNumberOfActiveDoFs());
+    ASSERT_EQ(8u, dofManager->getTotalNumberOfActiveDoFsWithoutGhost());
+    ASSERT_EQ(0u, dofManager->mapEqsID(varId, 0, 0));
+    ASSERT_EQ((size_t)-1, dofManager->mapEqsID(varId, 0, 8));
 
     delete dofManager;
 }
@@ -117,12 +117,12 @@ TEST(Discrete, DoF_ghost_nodes)
         dofManager->setNumberingType(DofNumberingType::BY_POINT);
         dofManager->construct();
 
-        ASSERT_EQ(1, dofManager->getNumberOfVariables());
-        ASSERT_EQ(9, dofManager->getTotalNumberOfActiveDoFs());
-        ASSERT_EQ(5, dofManager->getTotalNumberOfActiveDoFsWithoutGhost());
-        ASSERT_EQ(0, dofManager->mapEqsID(varId, 0, 0));
+        ASSERT_EQ(1u, dofManager->getNumberOfVariables());
+        ASSERT_EQ(9u, dofManager->getTotalNumberOfActiveDoFs());
+        ASSERT_EQ(5u, dofManager->getTotalNumberOfActiveDoFsWithoutGhost());
+        ASSERT_EQ(0u, dofManager->mapEqsID(varId, 0, 0));
         ASSERT_TRUE(dofManager->isGhostPoint(0, 8));
-        ASSERT_EQ(8, dofManager->mapEqsID(varId, 0, 8));
+        ASSERT_EQ(8u, dofManager->mapEqsID(varId, 0, 8));
 
         delete dofManager;
     }
@@ -135,12 +135,12 @@ TEST(Discrete, DoF_ghost_nodes)
         dofManager->setNumberingType(DofNumberingType::BY_POINT);
         dofManager->construct();
 
-        ASSERT_EQ(1, dofManager->getNumberOfVariables());
-        ASSERT_EQ(9, dofManager->getTotalNumberOfActiveDoFs());
-        ASSERT_EQ(4, dofManager->getTotalNumberOfActiveDoFsWithoutGhost());
+        ASSERT_EQ(1u, dofManager->getNumberOfVariables());
+        ASSERT_EQ(9u, dofManager->getTotalNumberOfActiveDoFs());
+        ASSERT_EQ(4u, dofManager->getTotalNumberOfActiveDoFsWithoutGhost());
         ASSERT_TRUE(dofManager->isGhostPoint(0, 0));
-        ASSERT_EQ(8, dofManager->mapEqsID(varId, 0, 4));
-        ASSERT_EQ(0, dofManager->mapEqsID(varId, 0, 5));
+        ASSERT_EQ(8u, dofManager->mapEqsID(varId, 0, 4));
+        ASSERT_EQ(0u, dofManager->mapEqsID(varId, 0, 5));
         delete dofManager;
     }
 }
